Returning the proper resource for the dome light texture.
