+++ src/hdmaya/adapters/aiSkydomeLightAdapter.cpp
@@ -34,6 +34,7 @@
 #include <hdmaya/adapters/lightAdapter.h>
 #include <hdmaya/adapters/mayaAttrs.h>
 #include <hdmaya/adapters/tokens.h>
+#include <hdmaya/utils.h>
 
 #include <maya/MPlugArray.h>
 
@@ -102,7 +103,15 @@ public:
         } else if (paramName == HdLightTokens->enableColorTemperature) {
             return VtValue(false);
         } else if (paramName == HdLightTokens->textureResource) {
-            return {};
+            auto fileObj =
+                GetConnectedFileNode(GetNode(), HdMayaAdapterTokens->color);
+            // TODO: Return a default, white texture?
+            // Ideally we would want to return a custom texture resource based
+            // on the color, but not sure how easy that would be.
+            if (fileObj == MObject::kNullObj) { return {}; }
+            return VtValue{GetFileTextureResource(
+                fileObj, GetFileTexturePath(MFnDependencyNode(fileObj)),
+                GetDelegate()->GetParams().textureMemoryPerTexture)};
         }
         return {};
     }
